Move files orphaned by the hack/generated removal to the right place
