+++ v2/internal/controller/controllers/recordings/TestOperatorModeWatchers.yaml
